refactor: Move AddTodoItem to @web/components/AddTodoItem/
